requirements.txt
certifi==2023.7.22
charset-normalizer==2.0.12
et-xmlfile==1.1.0
idna==3.4
numpy==1.25.2
openpyxl==3.1.2
pandas==1.3.3
Pillow==8.3.2
psutil==5.9.5
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
pytz==2023.3
qiskit==0.30.0
qiskit-aer==0.9.0
qiskit-aqua==0.9.5
qiskit-ibmq-provider==0.16.0
qiskit-ignis==0.6.0
requests==2.26.0
six==1.16.0
urllib3==1.26.16
